+++ rollup.config.js
@@ -1,6 +1,8 @@
 import svelte from 'rollup-plugin-svelte';
 import resolve from 'rollup-plugin-node-resolve';
 import css from 'rollup-plugin-css-only';
+import postcss from 'rollup-plugin-postcss'
+
 
 export default [
     {
@@ -17,7 +19,10 @@ export default [
                     dev: true,
                 },
             }),
-            css(),
+            postcss({
+                extract: true
+            }),
+            // css(),
             resolve(),
         ],
     },