+++ src/lib/MultiSelection.svelte
@@ -15,6 +15,35 @@
     }
 </script>
 
+{#each value as item, i}
+    <div
+        class="multiSelectItem {activeValue === i ? 'active' : ''} {isDisabled
+            ? 'disabled'
+            : ''}"
+        on:click={(event) =>
+            multiFullItemClearable ? handleClear(i, event) : {}}>
+        <div class="multiSelectItem_label">
+            {@html getSelectionLabel(item)}
+        </div>
+        {#if !isDisabled && !multiFullItemClearable}
+            <div
+                class="multiSelectItem_clear"
+                on:click={(event) => handleClear(i, event)}>
+                <svg
+                    width="100%"
+                    height="100%"
+                    viewBox="-2 -2 50 50"
+                    focusable="false"
+                    aria-hidden="true"
+                    role="presentation">
+                    <path
+                        d="M34.923,37.251L24,26.328L13.077,37.251L9.436,33.61l10.923-10.923L9.436,11.765l3.641-3.641L24,19.047L34.923,8.124 l3.641,3.641L27.641,22.688L38.564,33.61L34.923,37.251z" />
+                </svg>
+            </div>
+        {/if}
+    </div>
+{/each}
+
 <style>
     .multiSelectItem {
         background: var(--multiItemBG, #ebedef);
@@ -73,32 +102,3 @@
         vertical-align: top;
     }
 </style>
-
-{#each value as item, i}
-    <div
-        class="multiSelectItem {activeValue === i ? 'active' : ''} {isDisabled
-            ? 'disabled'
-            : ''}"
-        on:click={(event) =>
-            multiFullItemClearable ? handleClear(i, event) : {}}>
-        <div class="multiSelectItem_label">
-            {@html getSelectionLabel(item)}
-        </div>
-        {#if !isDisabled && !multiFullItemClearable}
-            <div
-                class="multiSelectItem_clear"
-                on:click={(event) => handleClear(i, event)}>
-                <svg
-                    width="100%"
-                    height="100%"
-                    viewBox="-2 -2 50 50"
-                    focusable="false"
-                    aria-hidden="true"
-                    role="presentation">
-                    <path
-                        d="M34.923,37.251L24,26.328L13.077,37.251L9.436,33.61l10.923-10.923L9.436,11.765l3.641-3.641L24,19.047L34.923,8.124 l3.641,3.641L27.641,22.688L38.564,33.61L34.923,37.251z" />
-                </svg>
-            </div>
-        {/if}
-    </div>
-{/each}