+++ src/lib/VirtualList.svelte
@@ -123,6 +123,23 @@
     });
 </script>
 
+<svelte-virtual-list-viewport
+    bind:this={viewport}
+    bind:offsetHeight={viewport_height}
+    on:scroll={handle_scroll}
+    style="height: {height};">
+    <svelte-virtual-list-contents
+        bind:this={contents}
+        style="padding-top: {top}px; padding-bottom: {bottom}px;">
+        {#each visible as row (row.index)}
+            <svelte-virtual-list-row>
+                <slot item={row.data} i={row.index} {hoverItemIndex}
+                    >Missing template</slot>
+            </svelte-virtual-list-row>
+        {/each}
+    </svelte-virtual-list-contents>
+</svelte-virtual-list-viewport>
+
 <style>
     svelte-virtual-list-viewport {
         position: relative;
@@ -140,19 +157,3 @@
         overflow: hidden;
     }
 </style>
-
-<svelte-virtual-list-viewport
-    bind:this={viewport}
-    bind:offsetHeight={viewport_height}
-    on:scroll={handle_scroll}
-    style="height: {height};">
-    <svelte-virtual-list-contents
-        bind:this={contents}
-        style="padding-top: {top}px; padding-bottom: {bottom}px;">
-        {#each visible as row (row.index)}
-            <svelte-virtual-list-row>
-                <slot item={row.data} i={row.index} {hoverItemIndex}>Missing template</slot>
-            </svelte-virtual-list-row>
-        {/each}
-    </svelte-virtual-list-contents>
-</svelte-virtual-list-viewport>